Make tracked model properties properly joinable.
Summary:
Release note: none

The internal properties of ObjectValue that are there for modeling the environment can be set or unset in different execution branches for various reasons. As a consequence they must be of a type that can be joined and the accessor properties must take joined values into account.

Restrictive properties, such as isPartial treat the join of true and false as true. Permissive properties, such as isSimple, treat the join of true and false as false.
Closes https://github.com/facebook/prepack/pull/1488

Differential Revision: D7057536

Pulled By: hermanventer

fbshipit-source-id: 9ef90e4087a892e08cc5f00a22053a6903311c41