Adding lazyObjectsRuntime option to the Prepack Website
Summary:
This addresses the issue https://github.com/facebook/prepack/issues/1282 although the issue also requests for --additionalFunctions option which is not part of this commit. User should now be able to see --lazyObjectsRuntime option in the options list on https://prepack.io/repl.html. Ideally this should be an enum but this will accept a freeform string in this revision.
Closes https://github.com/facebook/prepack/pull/1475

Differential Revision: D7059756

Pulled By: phanipattapu

fbshipit-source-id: 588f688a52e81d065cf19ac1fbd358a9332b81a0