Support for abstract interpretation of switches
Summary:
Thought I'd (re)start by just adding a hook point to the existing interpreter code to branch off in case of abstract interpretation. The two are likely to look quite different and keeping them separate will be tidier. Also added a basic test case to ensure that it fails when abstract interpretation is required for a switch.

Looking for feedback if this looks repulsive. If not, then I'll go make it do something basic as the next step (and add more tests) ;-)
Closes https://github.com/facebook/prepack/pull/1490

Differential Revision: D7070026

Pulled By: neilmacintosh

fbshipit-source-id: 0c10d0c9c80d1579ddbc128a32009b67f8c9a39b