Allowing specifying abstract function result type
Summary:
Release notes: None

In calls to __abstract, the first argument represents a type name or a template object.
Adding a new kind of type name: A colon followed by a simple type name.
This indicates that the abstract value is a function with the given known result type.

Also tweaked handling of Array concat so that it doesn't require knowing the exact elements
being concatenated (this is what motivated the abstract function result type change for me...)
Closes https://github.com/facebook/prepack/pull/1505

Differential Revision: D7102523

Pulled By: NTillmann

fbshipit-source-id: c8f81669eb938e75bde8231544ae42fe1788dbc9