Fix serialization ordering of ReactElement dependencies
Summary:
Release notes: none

As ReactElement's are immutable, we can't mutate them after we create them initially. This is problematic for serialization where properties in Prepack change over time. So instead, we emit a copy of the ReactElement shape at each point its needed as a dependency. This is also a tidy up of the ReactElement serialization too.
Closes https://github.com/facebook/prepack/pull/1502

Differential Revision: D7108327

Pulled By: trueadm

fbshipit-source-id: 88810f106c970241d26fc48bc118d78b6309594a