Handle React component render props
Summary:
Release notes: none

This PR is the work to get component render props working with the reconciler. We check if a render prop function is safe to use a new root for a component tree by scanning it's bindings using the leak ref traverser. If we find a binding that was created in the parent component, then it's not safe to make the render prop an additional function.
Closes https://github.com/facebook/prepack/pull/1478

Differential Revision: D7119766

Pulled By: trueadm

fbshipit-source-id: 8d286f1bc0ace009d2304643ba527b9932d2aa93