Distinguish leak from havoc
Summary:
Release note: Make terminology more standard

When an object is passed as an argument to an unknown function, or if it can be reached via another object that was passed to the function, then we talk about the object having leaked out of the heap that Prepack knows about into the environment that Prepack does not know about.

Provided that we can assume that the unknown function is generally well behaved, for example if we know that it will not modify global variables, or depend on global variables having values that were not set up during the initial global code execution, then we can allow the call to the unknown function without issuing a fatal error, provided that we forget everything that Prepack knows about the leaked object before the function call.

This process of forgetting about state because of unknown (uninterpreted) function calls is generally known has havocing the state. So, rather than talk about "leaked objects" and "leaked values" let's talk about havoced objects and havoced values. This is more in keeping with existing practice and decouples the concept fo forgetting state (havoc) from the concept of exposing state to external environment (leaking). This opens the door to scenarios where state may need to be havoced without there being a leak.

I've also added some todo comments to call out opportunities where it is not necessary to forget absolutely everything about an object.
Closes https://github.com/facebook/prepack/pull/1514

Differential Revision: D7125691

Pulled By: hermanventer

fbshipit-source-id: 9bb3ae6dded06af263a21bc76b3ccdecce08fbcc