Allow complex React class components to fold into simple class components
Summary:
Release notes: none

This PR introduces a big clean up and re-design of some of the React reconciler code to allow for better folding capabilities. This PR enables complex React class components to be folded into functional/simple React components. This involves cloning the class component data structure and copying it on to the functional component.

Finally, any original properties of the functional component, for example `getTrials` or `defaultProps` are copied from the original back over the top to ensure the existing semantics remain (otherwise `getTrials` etc will be missing from the root component).

This PR is a fundamental stepping stone to allow us to better fold component trees and also tidies up some of the existing logic to make it easier to reason with and understand.
Closes https://github.com/facebook/prepack/pull/1428

Differential Revision: D7146514

Pulled By: trueadm

fbshipit-source-id: 3c0581b4beb341dac1d563a8ec9861afa1bb1708