Remove react-rest console.error noise
Summary:
Release notes: none

We spam the console when running `test-react` and many of the errors are not errors but just noise. Interestingly this also brought about one test that genuinely is failing but before was bringing up an invalid bail-out with `dynamic-props.js`. I'll fix that as a follow up as it's unrelated to this PRs goal.
Closes https://github.com/facebook/prepack/pull/1529

Differential Revision: D7149640

Pulled By: trueadm

fbshipit-source-id: e3268bc92c62726a5686b1acfcfff4128b69f6cd