Add "node-react" compatibility mode
Summary:
Release notes: adds "node-react" compatibility mode

Adds the "node-react" compatibility mode, which is like "fb-www", but doesn't contain the internal FB mocks. This mode is ideal for library authors who might want to experiment with the features of the React compiler in an environment that has basic React/Relay mocks as well as `require` and `module.exports` partial support.

I'll be adding relevant documentation for this on the wiki too.
Closes https://github.com/facebook/prepack/pull/1528

Differential Revision: D7149660

Pulled By: trueadm

fbshipit-source-id: 3f00fa01bb881722b270393a27f817911c5a5256