Fix hasOwnProperty invariant failure
Summary:
Release notes: none

This fatals in the case we try to extract a descriptor from a topVal which can happen.

This lets the CallExpression fallback to leaving the call inact.

Unfortunately that havocs the arguments unnecessarily. So I also special cased it. I guess we should just make this a common pattern for pure built-ins.
Closes https://github.com/facebook/prepack/pull/1524

Reviewed By: trueadm

Differential Revision: D7152130

Pulled By: sebmarkbage

fbshipit-source-id: 24e5c02d36ae6a2e685b5b903c6cdfd0937b0834