Do not havoc frozen objects
Summary:
Release note: Do not havoc leaked objects if they are frozen

This still traverses the object and makes sure that any unfrozen locally created objects that are reachable from it are havoced.

sebmarkbage What is the rationale for leaving unfrozen objects unhavoced if they were not created in a pure context?
Closes https://github.com/facebook/prepack/pull/1520

Differential Revision: D7145807

Pulled By: hermanventer

fbshipit-source-id: aac41d3d814cb6cc9a0b486c3913a9a2f563458f