Attempt at abstract evaluation of switch cases using recursion.
Summary:
Adds abstract evaluation for switch cases. Effectively treats the switch case as a series of if-else statements to achieve the same outcome.

Currently does not support case blocks that result in an abrupt completion through `return/throw/continue`. That is work-to-go in the future.

A future PR will also expand on the tests to provide more coverage (and look for corner cases).
Closes https://github.com/facebook/prepack/pull/1519

Differential Revision: D7163238

Pulled By: neilmacintosh

fbshipit-source-id: 7ee824f90821d6a3d0d495ed01f8b09b0d190bbd