Allow simple object property access with an abstract dynamic key
Summary:
Release notes: Allow simple object computed property access with abstract property keys

The aim of this PR is to provide object computed property access when the property key is an abstract value and the object is a simple object. Furthermore, given the key is abstract, and converting it to a key might involve side-effects, this is only handled in a pure scope.
Closes https://github.com/facebook/prepack/pull/1530

Differential Revision: D7164200

Pulled By: trueadm

fbshipit-source-id: dfd8ac86985fd9d96fc18c6c28b954dff51eab27