Git rid of calls to Object.setPrototypeOf in the Prepack code base.
Summary:
Release note: Adapted Prepack source code to latest Flow

The latest version of Flow has a bug where it does not know about Object.setPrototypeOf. It turns out that we don't really need to call it anymore, so we may as well get rid of it.
Closes https://github.com/facebook/prepack/pull/1536

Differential Revision: D7178682

Pulled By: hermanventer

fbshipit-source-id: 5caca7796ff3f0aeab57ea8af4b0354d1e413061