Support additional function returning PossiblyNormalCompletion
Summary:
Release Notes: Allow additional functions to return PossiblyNormalCompletions.

Similar to the logic in `Program.js`, we need to emit an `if` statement at the end of an additional function that may return or throw.

In `functions.js` we process the PossiblyNormalCompletion, joining the consequent and alternate effects and creating a set of args + buildnode for emitting the conditional throw. Then where we would originally serialize/visit the return value, we will instead apply the joined effects and emit the conditional throw/return.
Closes https://github.com/facebook/prepack/pull/1523

Differential Revision: D7238882

Pulled By: cblappert

fbshipit-source-id: fd51bb4998394930960a70c8ef40b574e3244026