Protect against common firstRender issues
Summary:
Release notes: none

firstRender can fail when `Function.prototype.bind` is passed around because the firstRender removes the class component entirely, along with any concept of state etc. To prevent outputting the incorrect value, we now check for bind and if we're given something that is completely abstract as a child from a render, we bail out in that case too.
Closes https://github.com/facebook/prepack/pull/1573

Differential Revision: D7248829

Pulled By: trueadm

fbshipit-source-id: eebd7e6565352a37026928a20944377bb06129f5