Replace "sentinel Object" with "implicit conversion to object"
Summary:
Release note: Fix delayed conversions of abstract values to objects in pure mode

The "sentinel Object" wrapper name is a bit misleading, so I've renamed it to "implicit conversion to object".

The bug itself arises from the fact that the abstract value (wrapper) was stripped before it arrived at the serializer. This broke the dependency tracking logic of the serializer. The proper way to do this is to let the serializer do the stripping.

Unfortunately I do not have a small regression test case that gets fixed by this. I'm still looking, but meanwhile the bug fix will unblock Nikolai and perhaps others too.
Closes https://github.com/facebook/prepack/pull/1578

Differential Revision: D7257507

Pulled By: hermanventer

fbshipit-source-id: d2df04b6fb202deb50e46ec3a049c0feb60a35d8