Simplify truthy conditionals in condition context
Summary:
Release note: Simplify conditions like c ? {} : null to just c if used in a condition context

Resolves: #1582

The simplifier already simplifies c ? true : false to just c. It can also do this for truthy values that are known to convert to true and false, if the result of the conditional is used as a condition.
Closes https://github.com/facebook/prepack/pull/1583

Differential Revision: D7261552

Pulled By: hermanventer

fbshipit-source-id: cf2c16ffd1dd99a23ea9cf4f8065d0f495b5d749