Allow NewExpression constructor to recover in pure wrapper
Summary:
Release notes: none

It's common for a `new Something()` call to exist in pure wrappers where `Something` is an abstract value. Currently, this fails with a FatalError with no recovery. This PR adds a recover for this when in a pure wrapper so we can continue evaluating.
Closes https://github.com/facebook/prepack/pull/1590

Differential Revision: D7271876

Pulled By: trueadm

fbshipit-source-id: fc5192af6b356b7df9eb2f147bd8b53fe3790942