Use getProperty, setProperty and deleteProperty in React reconciliation
Summary:
Release notes: none

When we're reconciling the React component tree, we should circuit all objects that we know about (because we create them or manage them). By doing so, we emit less code to the generator/modified bindings and allow cases where we can modify/delete properties on so-called "frozen" objects like ReactElements and props (we do this for internal reasons, such as marking ReactElement's `ref` and `key` properties as deleted to avoid bail-out cases).
Closes https://github.com/facebook/prepack/pull/1605

Differential Revision: D7308534

Pulled By: trueadm

fbshipit-source-id: 3be367d18b634c448521ef06fa0dd050e68f6103