Make Path.pushRefinedContions linear
Summary:
Release note: Performance tweak

We now first simplify/refine all old path conditions before pushing them back onto the realm's list. This means that the length of the paths conditions used to refine the old conditions does not grow while doing the refinement, which removes a non linearity from this code path.

There should not be a significant reduction in precision, since each of the old conditions have already been simplified/refined in the light of all of the other old conditions before the latest condition gets pushed onto the condition list.

There is also a small bug fix for an invariant that should have been a conditional. This crept in during the update for Flow. Of course, a small regression test would have prevent this, but I have actually tried to construct such a test before and it proved to be harder than I have time for.
Closes https://github.com/facebook/prepack/pull/1617

Differential Revision: D7332414

Pulled By: hermanventer

fbshipit-source-id: c29707e9b4b1780d52706e23e18804c95802315a