Do not use variable before initialization
Summary:
Release note: Fix bug that caused variables to get used before their definitions have been processed.

Resolves: #1585

Following an code sequence that may terminate abruptly, or alternatively carry on normally, the generator that preceded the fork must be emitted into its parent and replaced with a new generator, before carrying on with the normal execution, since the normal execution is effectively nested inside the normal branch of the fork.

Note that this only occurs after the second such fork, which is why the bug was so elusive.
Closes https://github.com/facebook/prepack/pull/1619

Differential Revision: D7338429

Pulled By: hermanventer

fbshipit-source-id: e7463d87d20bde4f13bd6e69e0afe20c548f07c0