Provide a way to snapshot objects at temporal points.
Summary:
Release note: Allow subsequent mutating of Object.assign() arguments by snapshotting them

Fixes issue: #1480

The normal mode for Prepack is to run to the end of global code and then serialize the objects as they are at that point. This causes problems if objects are passed to functions that must be called at particular temporal points, since the calls must receive the objects as they are at that point, rather than as they are at the time when Prepack serializes them.

To solve this conundrum, there is now a way to get a snapshot of an object at a particular temporal point. This works by creating a new object and transferring the properties from the original object to the new object. Things are arranged so that the two objects are always referred to by the same name in the generated code. The nice end result is that subsequent modifications to the object does not affect the object that is passed to the function call where the snapshot was taken.

Following the snapshot, the original object has no own properties and must be havoced or marked simple partial, depending on the circumstances at the temporal point where the snapshot is taken.

The latter case arises, for example, when the target object of a call to Object.assign has been snapshotted because one of the other arguments is partial or fully abstract. Following the call, the object can still be used and all property references result in abstract values with runtime lookup behavior. It is also possible to assign new properties to the object and Prepack will track the values of those properties.

In order to deal with the fact that multiple compile time objects now correspond to a single runtime object, the first snapshot gets its own named temporal abstract value and the original object is associated with this object and the name of the latter becomes the name of the original. Once an object gets a name, all further writes to the object results in runtime assignments at the temporal point of the write.

Consequently subsequent snapshots just return this first snapshot, since the runtime object get incrementally updated. The property references are still cleared out, since snapshotting an object implies that it going to be modified in unspecified ways.
Closes https://github.com/facebook/prepack/pull/1563

Differential Revision: D7347283

Pulled By: hermanventer

fbshipit-source-id: 618873abf99bcadc4c74794aa7804a5ac5d04c12