Support inlining dynamic ReactElement "type" fields
Summary:
Release notes: none

This adds a really cool feature – dynamic ReactElement type. As per a conversation in Messenger yesterday by gaearon, it got me thinking about how we might "simplify" recursively on a conditional in complex cases and I decided it was probably worth adding to inline components. This PR makes it possible to inline component trees without having to break out when the type is an abstract conditional.
Closes https://github.com/facebook/prepack/pull/1647

Differential Revision: D7392497

Pulled By: trueadm

fbshipit-source-id: 170ae4829bd407410d8882c874144516cfd01190