Handle other receivers in AbstractObjectValue
It is fine to have the receiver be different and just delegate to the
underlying object.

If we're generating an get from an intrinsic object, that is simple, then
the receiver can't matter because there are no getters on it.

If the receiver on an abstract object value is self, then it is a simple
property access. We should delegate it to be a simple property access
on the inner element in that case.