Fix serialization of optimized functions that might throw exceptions.
Summary:
Release note: Stop generator entries from being serialized twice

Fixes issue: #1627

When the effects of optimized function is reduced to a generator, care must be taken to avoid duplicating the normal path of a function that could terminate with an exception.

It is a bit unfortunate that the effects has this duplication in there in the first place, but I was not able to figure out a way to avoid this and there are more pressing matters to attend to.
Closes https://github.com/facebook/prepack/pull/1670

Differential Revision: D7431202

Pulled By: hermanventer

fbshipit-source-id: 081713419c7cd46b77c76068fba89c62f49dea70