We can safely handle other receivers than self
We can assume that if GetPartial gets called on an object with a different
receiver than itself, that means that there were no properties further
up the chain.

This should be something that we can preserve in the future, but currently
it relies on the bug in 1675 which ignores the prototype chain.