Add rule to help simplify expressions with conditions involving == and !=
Summary:
Release note: Simplification now works better when == and != are used.

The main goal of this PR is simplify away unnecessary object creations such as the one in the test case.

Working on this, however, I ran into cases where tweaks to the simplifier would cause "push false" invariant failures. This time around I could not find a way to avoid it, so the time has come to exploit it. Consequently, the biggest part of this PR is to recover when we detect a contradiction by reverting to behavior we would have had, had we known that assuming the condition is true (or false) is actually invalid.

While there is a lot of code, the transformations are pretty simple.
Closes https://github.com/facebook/prepack/pull/1679

Differential Revision: D7452855

Pulled By: hermanventer

fbshipit-source-id: 507f0a1a6dbe9432ad5654e488214b8da7f7197c