Push path conditions when simplifying conditional expressions.
Summary:
Release note: none

The basic change here is to use path conditions when simplifying conditional expressions. Doing so led to an explosion of work in the simplifier, so I had to introduce more flags to reign in the simplifier and preventing it from trying to simplify things that have already been simplified.
Closes https://github.com/facebook/prepack/pull/1684

Differential Revision: D7457002

Pulled By: hermanventer

fbshipit-source-id: 30ecde01c23b52871cdf43b24bf9ab9c5eecff4f