Various Fixes to GetPartial
Summary:
This adds havocing in the right places that was not done properly in the previous solution.

This also allows non-self Receivers. For simple objects, this is allowed because they can't possibly have any meaningful impact if the object is truly simple since only getters expose the receiver concept.

More than that we can also make an assumption for non-simple objects, in pure mode. I assume that if the receiver is not self then the only reason we got here is because we concluded that anything else on the prototype chain cannot have had this property key on it. For example, that happens with a ToObject coercion. The original values doesn't have any properties on it.

A fix to #1675 might need to account for this scenario because currently we never propagate up the prototype chain so this can never happen.

However, even if there might be something else shadowing this on the prototype chain, this property access will still be ok since the residual is serialized as a property access on the receiver - which is havoced.

Fixes #1676
Closes https://github.com/facebook/prepack/pull/1680

Differential Revision: D7460594

Pulled By: sebmarkbage

fbshipit-source-id: e71dee4a6edf9b8a32a09016e7ac455ad03ad792