Emit explicit conversion to object
Summary:
Release note: Fix problem where the results of To.ToObject were lost during serialization

Fixes issue: #1595

The serializer will now serialize the abstract result of calling To.ToObjectPartial on an abstract value as "Object.assign(abstract_value)". In the one case where the serialized code is known to implicitly convert the abstract value to an object anyway, the caller of To.ToObjectPartial strips away the wrapper in order to avoid the redundant explicit conversion.

Since I ended up looking at every call of To.ToObjectPartial, I used this an opportunity to merge it into To.ToObject.

I also made AbstractObjectValue.$GetPrototypeOf aware of the wrapper and made it return concrete values if the type of the wrapped primitive is known.
Closes https://github.com/facebook/prepack/pull/1690

Differential Revision: D7490354

Pulled By: hermanventer

fbshipit-source-id: 75146883d8b949c4e3e7848bd2e030f7b7c4d147