Disable some simplifier special cases when isCondition is true.
Summary:
Release note: none

If the x and y part of a c ? x : y are simplified under the assumption that they will only be used contexts where they are converted to Booleans, then it is not safe to use them in equality comparisons. Hence the simplification rules that do so must be disabled in this situation.
Closes https://github.com/facebook/prepack/pull/1698

Differential Revision: D7511274

Pulled By: hermanventer

fbshipit-source-id: f0329c00523cb3c06de9e0dde3ab55104dbb07c7