Clean up Widen
Summary:
Release note: none

widenValues had some funny looking code for the case where one of its parameters are void. See issue #1554.

While the code seems wrong, it is also not clear how to write a test case that will fail because of it, so instead I've tightened up the type annotations to prevent a void argument to widenValues. That leaves the funky decision of what to do about undefined values to _widenArrayOfsMapEntries, where it is at least a bit more contextual. This is also untested because this method is more a place holder right now than a part of a tested scenario.
Closes https://github.com/facebook/prepack/pull/1711

Differential Revision: D7526508

Pulled By: hermanventer

fbshipit-source-id: 8dc5bb5f00159f55eaac8f4908563405c1332c7b