Make AbstractValue.describe more linear and avoid calling it
Summary:
Release note: none

AbstractValue.describe can get very expensive when the same (expensive to describe) condition shows up many times inside a large conditional value. This can be avoided with a suitable occurs check.

Of course, if the description is going to be ignored, there is not much point in generating it in the first place, so there is now a flag to prevent that.
Closes https://github.com/facebook/prepack/pull/1712

Differential Revision: D7528371

Pulled By: hermanventer

fbshipit-source-id: e6b4579f5a0b843d6db437dcb70746f1df972c6b