Fixes an infinite loop problem in reconciler
Summary:
Release notes: none

From the UFI, I noticed that entered an infinite loop because we keep adding the same branches to the branches array – `branchedComponentTrees`, even though they were already processed somewhere else in the app. We need to check if we evaluated the component before we add it to `branchedComponentTrees`.
Closes https://github.com/facebook/prepack/pull/1714

Differential Revision: D7534892

Pulled By: trueadm

fbshipit-source-id: b8d636101eab9c696635836c5d22472a5c85a47c