Make AbstractValue.createFromType take arguments
Summary:
Release note: none

AbstractValue.createFromType assumed that the values it produces will have no arguments, not least because of the expectation that arguments must be fed into the build function in all cases.

Since then it has become convenient to create some values that take no build function but have arguments. These have been constructed by using AbstractValue.createFromType and then adding arguments afterwards, which is an anti-pattern because of hashing concerns and general code hygiene.

Given current reality, the constructor of AbstractValue no longer sports an invariant that requires a build function to be present if arguments are supplied, and createFromType now has an extra parameter for arguments.
Closes https://github.com/facebook/prepack/pull/1719

Differential Revision: D7553960

Pulled By: hermanventer

fbshipit-source-id: db6afe615582db2086e7b1a8a048cee4f198f710