Treat fbt as a global object
Summary:
It seems like it's unsafe to rename on www (it breaks Haste) so we should add it to the whitelist of globals.

It's both a function *and* it has properties that are functions, but for now I'll model it as an object since we never use it directly. In fact it seems like www transforms direct calls on it to `fbt._` calls or something like that so maybe keeping it as an object is right. We don't have a single direct call to `fbt()` as a function in our bundle.

This fix unblocks testing the bundle initialization in Jest www environment.
Closes https://github.com/facebook/prepack/pull/1725

Reviewed By: trueadm

Differential Revision: D7568551

Pulled By: gaearon

fbshipit-source-id: bdfc7c949d6be698e1f58a85c1e39f8b6d5b6bae