Handle cases where gDSFP is given an abstract state
Summary:
Release notes: none

When working on internal projects, I found cases where the `getDerivedStateFromProps` would throw a FatalError because the partial state wasn't an abstract object value or object value. In this case we can see that the abstract value is a conditional and handle the paths to ensure we get the correct results without failing.
Closes https://github.com/facebook/prepack/pull/1737

Differential Revision: D7615094

Pulled By: trueadm

fbshipit-source-id: 5c868a46936a108c277ac52c2675feabc1f4df8c