Join JoinedAbruptCompletions when optimizing functions.
Summary:
Release note: none

The effects constructed for an optimized function must join all paths that may leave the function in order to provide a full accounting of what the function might do to the global environment. Without this, the global environment can get polluted by side effects in the abrupt branches, because these effects are applied to the environment when constructing the generator from the overall effects, but are not reversed when backing out the overall effects.

This can break things in many different ways, one of which is illustrated in the modification of return-or-throw-simple.js, where the last return x is serialized as undefined rather than the abstract value, because x got polluted during the construction of the effects for func1.
Closes https://github.com/facebook/prepack/pull/1749

Differential Revision: D7664313

Pulled By: hermanventer

fbshipit-source-id: 3b061d16a4f38918f461b792b660954332ae6fa7