Preserve order of joined throw completions at function end
Summary:
Release note: none

The code for disentangling return completions from throw completions uses a recursive algorithm that caused the throw completions to get reversed. Fix this to prepend rather than append so that the order comes out right.

This is also an opportunity to not mess with the path conditions.
Closes https://github.com/facebook/prepack/pull/1752

Differential Revision: D7676676

Pulled By: hermanventer

fbshipit-source-id: f10326e21bdf3f15c2ddf7eac02ef02cf8d155d3