Make effects into a class in smallest possible change
Summary:
Release Notes: None

There are various reasons we want `Effects` to be a class. It will allow for better encapsulation as well as testing of some invariants. This is the simplest/most mechanical way I could think of to transform `Effects` into a class. Subsequent refactors should make the class more than a wrapper for the original array.
Closes https://github.com/facebook/prepack/pull/1754

Differential Revision: D7678281

Pulled By: cblappert

fbshipit-source-id: 90690bdd2768cfa4a6b99a8884e807a8b6c74f96