Use actual location as cache key if available
Summary:
Release note: none

I noticed that the stack properties for Error objects were being set to the same value even when the objects were constructed in separate lines.

It turns out that some recent caching logic used the containing function as key and thus unified all error location descriptions encountered in the same function to a single value. So part of this PR is just a fix that uses the actual location, if available, as the cache key.

This unmasked another recently introduced problem: There are now invariant checks that happen if you get a property from the global object. The function where the checks are triggered is called as part of the describe logic. This logic was largely disabled because of the caching bug, so the checks did not fire in any test case and we missed out on the fact that the check don't work if the property get logic is not part of the source program.

To fix that, all calls to $Get that are done for internal purposes have now been replaced by calls to _SafeGetDataPropertyValue.
Closes https://github.com/facebook/prepack/pull/1755

Differential Revision: D7683687

Pulled By: hermanventer

fbshipit-source-id: 1c811fffc6ecf275f5fabd75afbd9e8cfc051247